+++ app/lib/pages/login_page.dart
@@ -1,13 +1,11 @@
 // ignore_for_file: prefer_const_constructors
 
-import 'package:app/components/LoginPage/login_button.dart';
-import 'package:app/components/LoginPage/register_button.dart';
+import 'package:app/components/login_page/register_button.dart';
 import 'package:app/components/custom_text_form_field.dart';
 import 'package:app/components/logo.dart';
 import 'package:app/constants/colors.dart';
 import 'package:app/constants/size.dart';
-import 'package:app/navigation/routes.dart';
-import 'package:app/object/user_controller.dart';
+import 'package:app/controller/user_controller.dart';
 import 'package:app/screens/home_screen.dart';
 import 'package:flutter/material.dart';
