+++ app/lib/pages/register_page.dart
@@ -1,16 +1,12 @@
 // ignore_for_file: prefer_const_constructors
 
-import 'dart:convert';
-import 'package:app/components/RegisterPage/make_account_button.dart';
 import 'package:app/components/custom_text_form_field.dart';
 import 'package:app/components/logo.dart';
 import 'package:app/constants/colors.dart';
 import 'package:app/constants/size.dart';
-import 'package:app/object/user_controller.dart';
+import 'package:app/controller/user_controller.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
-import 'package:app/constants/colors.dart';
-import 'package:http/http.dart' as http;
 
 class RegisterPage extends StatefulWidget {
